Add recipe folder, baguette.html and update index
As part of the second iteration of the curriculum a recipes folder was
created and baguette.html was added as the first recipe. index.html was
updated to contain a link to the new recipe.
Of note that this commit does NOT include a typo fix in README. As
stated in my previous commit, I'd like to not jumble typo and code
commits together and am trying to see if this is a viable way to do so.