Add more recipes and return link
As part of the fourth and last iteration of the curriculum, two more
recipes have been added: Pizza Boston and Fish Bordelaise. A lot less
effort went into these but the practice was still there.
Also added a "Back to Recipe Page" link to each recipe, including
Baguette that returns to the main page. Was not demanded by the
assignment but felt right.